refactor(room-list): Restructure room list logic and improve code style
- Moved logic from hooks to utils/grpc-helpers
- Removed overly concise functions
- Enhanced overall code organization and readability
