+++ src/components/room-list/RoomListCard.tsx
@@ -1,12 +1,12 @@
 import { Button, Heading, HStack } from "@chakra-ui/react";
 
-interface roomListDataProps {
+interface RoomListDataProps {
     name: string;
     id: bigint;
     handleIntoRoom: (id: bigint) => void;
 }
 
-export const RoomListCard: React.FC<roomListDataProps> = ({ name, id , handleIntoRoom}) => {
+export const RoomListCard: React.FC<RoomListDataProps> = ({ name, id , handleIntoRoom}) => {
   return (
     <Button radioGroup="xl" height={"100px"} width={"100%"} variant={"ghost"} onClick={() => handleIntoRoom(id)} >
       <HStack direction={"row"}>
