+++ src/utils/storage/user-storage.ts
@@ -24,6 +24,17 @@ export const storage = {
     }  
     return { userId: BigInt(data.userId), tokenPair: data.tokenPair };
   },
+
+  getUserId: (): bigint | null => {
+    const userAuth = storage.getUserAuth();
+    return userAuth ? userAuth.userId : null;
+  },
+
+  getTokenPair: (): TokenPair | null => {
+    const userAuth = storage.getUserAuth();
+    return userAuth ? userAuth.tokenPair : null;
+  },
+
   clearUserAuth: () => {
     localStorage.removeItem("userAuth");
   },
@@ -35,7 +46,5 @@ export const storage = {
     } catch (error) {
       console.error("save token error: ", error);
     }
-    
-
   }
 };
